# api/openapi.yaml
openapi: 3.0.0
info:
  description: This is a sample User management server based on the OpenAPI 3.0 specification.
  title: Swagger user management service - OpenAPI 3.0
  version: 1.0.0
servers:
- url: https://user-management.swagger.io/api/v1
tags:
- description: Everything about users
  name: users
- description: Everything about roles
  name: roles
- description: Everything about user-roles
  name: user-roles
paths:
  /users:
    get:
      operationId: getAllUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/user'
                type: array
          description: successful operation
        "404":
          description: Users not found
      summary: Provides a list of users
      tags:
      - users
  /users/{id}:
    delete:
      description: delete user
      operationId: deleteUserByID
      parameters:
      - description: ID of user deleting
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not deleting
      summary: delete user by ID
      tags:
      - users
    get:
      description: Returns a single User
      operationId: getUserById
      parameters:
      - description: ID of user to returns
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
      summary: Find user by ID
      tags:
      - users
    put:
      description: Returns a change user
      operationId: editUserByID
      parameters:
      - description: ID of user changing
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        description: Update an existent user
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not changing
      summary: edit user by ID
      tags:
      - users
  /roles:
    get:
      operationId: getAllRoles
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/role'
                type: array
          description: successful operation
        "404":
          description: Users not found
      summary: Provides a list of roles
      tags:
      - roles
  /roles/{id}:
    delete:
      description: delete role
      operationId: deleteRoleByID
      parameters:
      - description: ID of role deleting
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Role not deleting
      summary: delete role by ID
      tags:
      - roles
    get:
      description: Returns a single role
      operationId: getRoleById
      parameters:
      - description: ID of role to returns
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/role'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Role not found
      summary: Find role by ID
      tags:
      - roles
    put:
      description: Returns a change role
      operationId: editRoleByID
      parameters:
      - description: ID of role changing
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/role'
        description: Update an existent role
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/role'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Role not changing
      summary: edit role by ID
      tags:
      - roles
  /user-roles/{userid}:
    get:
      description: Returns user roles
      operationId: getUserRolesByUserId
      parameters:
      - description: ID of user to returns
        explode: false
        in: path
        name: userid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/user_role'
                type: array
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User roles not found
      summary: Find user-roles by userID
      tags:
      - user-roles
    put:
      description: Returns a change user roles
      operationId: editUserRolesByUserID
      parameters:
      - description: ID of user changing
        explode: false
        in: path
        name: userid
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/user_role'
              type: array
        description: Update an existent user roles
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User role not changing
      summary: edit user role by userID
      tags:
      - user-roles
components:
  schemas:
    user:
      example:
        Active: true
        UserName: Ivanov.I
        Email: ivanov_i@example.com
        FirstName: Ivan
        FullName: Ivan Ivanov
        UserGroupId: 1
        Id: f7dabcd1-df8b-4fe6-9666-bf123782593e
        LastName: Ivanov
        Deleted: false
      properties:
        Id:
          example: f7dabcd1-df8b-4fe6-9666-bf123782593e
          type: string
        UserName:
          example: Ivanov.I
          type: string
        FirstName:
          example: Ivan
          type: string
        LastName:
          example: Ivanov
          type: string
        FullName:
          example: Ivan Ivanov
          type: string
        Email:
          example: ivanov_i@example.com
          type: string
        UserGroupId:
          example: 1
          format: int64
          type: integer
        Active:
          example: true
          type: boolean
        Deleted:
          example: false
          type: boolean
      type: object
    role:
      example:
        activity: true
        name: viewing the contract
        reading: false
        id: 1dr456h8-df8b-4fe6-9666-1aq5we9rt2y6
        changing: false
        viewing: true
      properties:
        id:
          example: 1dr456h8-df8b-4fe6-9666-1aq5we9rt2y6
          type: string
        name:
          example: viewing the contract
          type: string
        reading:
          example: false
          type: boolean
        changing:
          example: false
          type: boolean
        viewing:
          example: true
          type: boolean
        activity:
          example: true
          type: boolean
      type: object
    user_role:
      example:
        roleId: 1dr456h8-df8b-4fe6-9666-1aq5we9rt2y6
        active: true
        userId: f7dabcd1-df8b-4fe6-9666-bf123782593e
      properties:
        userId:
          example: f7dabcd1-df8b-4fe6-9666-bf123782593e
          type: string
        roleId:
          example: 1dr456h8-df8b-4fe6-9666-1aq5we9rt2y6
          type: string
        active:
          example: true
          type: boolean
      type: object